refactor into testable functinos
