+++ packages/node-build-scripts/src/constants.mjs
