+++ packages/node-build-scripts/src/sass/sassSvgInliner.mjs
