+++ packages/node-build-scripts/src/svg/svgOptimizer.mjs
