+++ packages/node-build-scripts/src/utils.mjs
